pyproject.toml
[tool.poetry]
name = "comforticons"
version = "0.1.0"
description = "ComfortIcons is an identicon generation module for different identicon/icon/gravatar/avatar flavors"
authors = ["blackwarlow"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/blackwarlow/comforticons"
documentation = "https://github.com/blackwarlow/comforticons"
keywords = ["identicon", "avatar", "gravatar"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.5.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
vermin = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
